+++ Libirary_/src/main/java/com/example/libirary_/AddBookSceneController.java
@@ -1,57 +1,19 @@
 package com.example.libirary_;
 
+import AdminPackage.AddBookFunctions;
+import AdminPackage.AdminFunctionalities;
+import InterfacesPackage.CommonFunctions;
+import com.example.libirary_.Book;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.scene.control.Button;
 import javafx.scene.control.CheckBox;
 import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
 
-public class AddBookSceneController implements CommonFunctions {
-    @FXML
-    private TextField nameTextField,authorTextField,genreTextField,publishyearTextField,imagepathTextField,priceTextField;
-    @FXML
-    private CheckBox statusCheckBox;
-    @FXML
-    private Label successLabel;
-    public void  CheckIfInputIsValid(){
-        try {
-            if(nameTextField.getText().isEmpty()
-                    ||authorTextField.getText().isEmpty()
-                    ||genreTextField.getText().isEmpty()
-                    ||publishyearTextField.getText().isEmpty()
-                    || imagepathTextField.getText().isEmpty()
-                    ||priceTextField.getText().isEmpty()){
-                showAlert("Please Enter all the data");
-                successLabel.setText("");
-                return;
-            }
-        }catch (Exception e){
-            showAlert("An error happened");
-            System.out.println(e.getMessage());
-        }
+import java.io.IOException;
 
-
-    }
-    public void GetData(){
-        String name = nameTextField.getText()
-                ,author =authorTextField.getText()
-                ,genre= genreTextField.getText()
-                ,imagepath=imagepathTextField.getText(),status;
-        int publishyear=Integer.parseInt(publishyearTextField.getText());
-        float price=Float.parseFloat(priceTextField.getText());
-        if(statusCheckBox.isSelected()){
-            status="In Stock";
-        }
-        else {
-            status = "Out of Stock";
-        }
-        successLabel.setText("Book added successfully");
-        Book newBook = new Book(name,author,status,publishyear,price,genre,imagepath);
-        System.out.println(Book.books.size());
-    }
-    public void AddBook(ActionEvent event){
-        CheckIfInputIsValid();
-        GetData();
+public class AddBookSceneController extends AddBookFunctions implements CommonFunctions {
+    public void Back(ActionEvent event) throws IOException {
+        BackToAdmin(event);
     }
 }
