+++ Libirary_/src/main/java/com/example/libirary_/User.java
@@ -6,7 +6,7 @@ public class User {
     static int userCounter =0;
     UserInformation userInfo;
     static User CurrentUser;
-   static ArrayList<User> users = new ArrayList<User>();
+   public static ArrayList<User> users = new ArrayList<User>();
 
     public User(String userName, String email, String password) {
         userCounter++;
