+++ Libirary_/src/main/resources/com/example/libirary_/UpdateBookInfo.fxml
@@ -10,9 +10,9 @@
    <children>
       <TextField fx:id="indexsearchTextField" layoutX="45.0" layoutY="31.0" promptText="Book Title" />
       <Button fx:id="searchButton" layoutX="229.0" layoutY="31.0" mnemonicParsing="false" onAction="#ViewBookToUpdate" text="Show" />
-      <TextField fx:id="updateAuthor" layoutX="419.0" layoutY="31.0" onAction="#setUpdate2" />
+      <TextField fx:id="updateAuthor" layoutX="419.0" layoutY="31.0" />
       <Label fx:id="update2" layoutX="369.0" layoutY="35.0" text="Author" />
-      <TextField fx:id="updatePrice" layoutX="420.0" layoutY="83.0" onAction="#setUpdatePrice" />
+      <TextField fx:id="updatePrice" layoutX="420.0" layoutY="83.0" />
       <TextField fx:id="UpdateTitle" layoutX="419.0" layoutY="200.0" />
       <TextField fx:id="genreTextField" layoutX="420.0" layoutY="309.0" />
       <TextField fx:id="UpdatePublishYear" layoutX="420.0" layoutY="249.0" />
@@ -22,8 +22,12 @@
       <Label fx:id="Update_PublishYear" layoutX="348.0" layoutY="253.0" text="Publish Year" />
       <Button fx:id="updateButton" layoutX="255.0" layoutY="351.0" mnemonicParsing="false" onAction="#UpdateBookInfo" text="Update" />
       <Label layoutX="367.0" layoutY="313.0" text="Genre" />
-      <CheckBox fx:id="statusCheckBox" layoutX="435.0" layoutY="147.0" mnemonicParsing="false" text="In Stock" />
+      <CheckBox fx:id="statusCheckBox" layoutX="431.0" layoutY="130.0" mnemonicParsing="false" prefHeight="0.0" prefWidth="64.0" text="In Stock" />
       <TextField fx:id="authorTextField" layoutX="45.0" layoutY="75.0" promptText="Book Author" />
       <Label fx:id="successLabel" layoutX="52.0" layoutY="301.0" />
+      <CheckBox fx:id="availableCheckBox" layoutX="431.0" layoutY="156.0" mnemonicParsing="false" text="Available For Borrow" />
+      <TextField fx:id="expiredatwTextField" layoutX="419.0" layoutY="351.0" />
+      <Label layoutX="350.0" layoutY="355.0" text="Expire Date" />
+      <Button layoutX="94.0" layoutY="351.0" mnemonicParsing="false" onAction="#Back" text="Back" />
    </children>
 </AnchorPane>
